Rework maintainer documentation
Tips about setting up repositories are removed because being a
proficient contributor is kind of a requirement, and new maintainers
probably do not need to be taught how to setup a git
repository, or how to add remotes.
References to master, or to the merge down workflow are dropped as well.