Make addressing deprecations acceptable in a patch release
Addressing a deprecation is not a bugfix, it does not make the software
more stable. Deprecations are annoying though, and unless we switch to
using the ~ operator instead of the ^ operator for our dependencies, we
might get new ones out of the blue.
I propose to explicitly allow PRs that address deprecations in patch
releases, so that users do not have to wait until the next minor to have
fewer deprecations.