Improve menu generation
So the origin of the menu's not working is in the way the menu's are defined. The doctrine parser did use * as a deep search wildcard which is not according to the behavior sphinx does it. As the toctree is a directive defined by sphinx we try to follow that one closely. The correct way of writing a toctree is

```
.. toctree::
   :glob:

   **
```

This will include all documents and sections in a project. starting at the level of the current page.